Refactored the parsing of examples as expectations
